+++ KarhooUISDK/Screens/CheckoutScreen/Checkout/VehicleDetailsCard.swift
@@ -34,8 +34,6 @@ struct VehicleDetailsCard: View {
                     Text(viewModel.fleetName)
                         .font(Font(KarhooUI.fonts.captionSemibold()))
                         .foregroundColor(KarhooUI.colors.text.getColor())
-                    
-                    
                 }
                 
             }
@@ -69,40 +67,4 @@ struct VehicleDetailsCardViewModel {
     let fleetName: String
     let carIconUrl: String
     let fleetIconUrl: String
-}
-
-struct NewVehicleCapacity: View {
-    var passangerCapacity: Int
-    var luggageCapacity: Int
-    var body: some View {
-        HStack {
-            CapacityCard(iconName: "kh_uisdk_passenger_capacity_icon", value: passangerCapacity)
-            CapacityCard(iconName: "kh_uisdk_luggage_icon", value: luggageCapacity)
-        }
-    }
-    
-    struct CapacityCard: View {
-        var iconName: String
-        var value: Int
-        var body: some View {
-            HStack {
-                Group {
-                    Image(uiImage: .uisdkImage(iconName))
-                        .frame(
-                            width: UIConstants.Dimension.Icon.small,
-                            height: UIConstants.Dimension.Icon.small
-                        )
-                    Text("\(value)")
-                        .font(Font(KarhooUI.fonts.footnoteSemiold()))
-                        .foregroundColor(KarhooUI.colors.text.getColor())
-                }
-            }
-            .padding(UIConstants.Spacing.xSmall)
-            .background(
-                RoundedRectangle(cornerRadius: UIConstants.CornerRadius.xSmall)
-                    .fill(KarhooUI.colors.background1.getColor())
-            )
-        }
-    }
-    
 }
